+++ capa/features/extractors/drakvuf/helpers.py
@@ -29,8 +29,8 @@ def index_calls(report: DrakvufReport) -> dict[ProcessAddress, dict[ThreadAddres
             # we ignore the pid 0 since it's a system process and it's unlikely for it to
             # be hijacked or so on, in addition to capa addresses not supporting null pids
             continue
-        proc_addr = ProcessAddress(pid=call.pid, ppid=call.ppid)
-        thread_addr = ThreadAddress(process=proc_addr, tid=call.tid)
+        proc_addr = ProcessAddress(pid=call.pid, ppid=call.ppid, id=0)
+        thread_addr = ThreadAddress(process=proc_addr, tid=call.tid, id=0)
         if proc_addr not in result:
             result[proc_addr] = {}
         if thread_addr not in result[proc_addr]: